// src/app/core/model/corporate-account.ts
export interface ICorporateAccount {
    _id: string;
    corporateName: string;
    address: string;
    imageURL: string;
    contractImageURL: string;
    email: string;
    phone: string;
    password: string;
    accessToken: string;
    tripRate: number;
    contactPerson: string;
    contactPersonPhone: string;
}